Finished BookRoom.cs and tests

+++ Model/Room.cs
@@ -5,7 +5,7 @@ public class Room
 {
     public long Id { get; set; }
     public int Number { get; set; }
-    public int HotelId { get; set; }
+    public long HotelId { get; set; }
     public Hotel Hotel { get; set; }
     public virtual ICollection<Booking> Bookings { get; set; }
 }
